requirements.txt
PyMuPDF
langchain-google-genai
pillow
pytesseract
pdf2image
opencv-python
pandas